+++ keba_keenergy_api/constants.py
@@ -43,67 +43,71 @@ HEAT_CIRCUIT_PREFIX: Final[str] = "APPL.CtrlAppl.sParam.heatCircuit"
 class ControlValue(NamedTuple):
     value: str
     data_type: type[float | int | str]
-    read_only: bool
+    read_only: bool = True
+    human_readable: type[IntEnum] | None = None
 
 
 class Options(Enum):
-    HOT_WATER_TANK_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHotWaterTanks", int, read_only=True)
-    HEAT_PUMP_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHeatPumps", int, read_only=True)
-    HEAT_CIRCUIT_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHeatingCircuits", int, read_only=True)
+    HOT_WATER_TANK_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHotWaterTanks", int)
+    HEAT_PUMP_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHeatPumps", int)
+    HEAT_CIRCUIT_NUMBERS: Final[ControlValue] = ControlValue("systemNumberOfHeatingCircuits", int)
 
 
 class Outdoor(Enum):
-    TEMPERATURE: Final[ControlValue] = ControlValue("values.actValue", float, read_only=True)
+    TEMPERATURE: Final[ControlValue] = ControlValue("values.actValue", float)
 
 
 class HotWaterTank(Enum):
-    TEMPERATURE: Final[ControlValue] = ControlValue("topTemp.values.actValue", float, read_only=True)
-    OPERATING_MODE: Final[ControlValue] = ControlValue("param.operatingMode", int, read_only=False)
-    MIN_TEMPERATURE: Final[ControlValue] = ControlValue("param.reducedSetTempMax.value", float, read_only=False)
+    TEMPERATURE: Final[ControlValue] = ControlValue("topTemp.values.actValue", float)
+    OPERATING_MODE: Final[ControlValue] = ControlValue(
+        "param.operatingMode",
+        int,
+        read_only=False,
+        human_readable=HotWaterTankOperatingMode,
+    )
+    MIN_TEMPERATURE: Final[ControlValue] = ControlValue(
+        "param.reducedSetTempMax.value",
+        float,
+        read_only=False,
+    )
     MAX_TEMPERATURE: Final[ControlValue] = ControlValue("param.normalSetTempMax.value", float, read_only=False)
 
 
 class HeatPump(Enum):
-    NAME: Final[ControlValue] = ControlValue("param.name", str, read_only=True)
-    STATUS: Final[ControlValue] = ControlValue("values.heatpumpState", int, read_only=False)
-    CIRCULATION_PUMP: Final[ControlValue] = ControlValue("CircPump.values.setValueScaled", float, read_only=True)
-    INFLOW_TEMPERATURE: Final[ControlValue] = ControlValue("TempHeatFlow.values.actValue", float, read_only=True)
-    REFLUX_TEMPERATURE: Final[ControlValue] = ControlValue("TempHeatReflux.values.actValue", float, read_only=True)
-    SOURCE_INPUT_TEMPERATURE: Final[ControlValue] = ControlValue("TempSourceIn.values.actValue", float, read_only=True)
-    SOURCE_OUTPUT_TEMPERATURE: Final[ControlValue] = ControlValue(
-        "TempSourceOut.values.actValue",
-        float,
-        read_only=True,
+    NAME: Final[ControlValue] = ControlValue("param.name", str)
+    STATUS: Final[ControlValue] = ControlValue(
+        "values.heatpumpState",
+        int,
+        read_only=False,
+        human_readable=HeatPumpStatus,
     )
-    COMPRESSOR_INPUT_TEMPERATURE: Final[ControlValue] = ControlValue(
-        "TempCompressorIn.values.actValue",
-        float,
-        read_only=True,
-    )
-    COMPRESSOR_OUTPUT_TEMPERATURE: Final[ControlValue] = ControlValue(
-        "TempCompressorOut.values.actValue",
-        float,
-        read_only=True,
-    )
-    COMPRESSOR: Final[ControlValue] = ControlValue("Compressor.values.setValueScaled", float, read_only=True)
-    HIGH_PRESSURE: Final[ControlValue] = ControlValue("HighPressure.values.actValue", float, read_only=True)
-    LOW_PRESSURE: Final[ControlValue] = ControlValue("LowPressure.values.actValue", float, read_only=True)
+    CIRCULATION_PUMP: Final[ControlValue] = ControlValue("CircPump.values.setValueScaled", float)
+    INFLOW_TEMPERATURE: Final[ControlValue] = ControlValue("TempHeatFlow.values.actValue", float)
+    REFLUX_TEMPERATURE: Final[ControlValue] = ControlValue("TempHeatReflux.values.actValue", float)
+    SOURCE_INPUT_TEMPERATURE: Final[ControlValue] = ControlValue("TempSourceIn.values.actValue", float)
+    SOURCE_OUTPUT_TEMPERATURE: Final[ControlValue] = ControlValue("TempSourceOut.values.actValue", float)
+    COMPRESSOR_INPUT_TEMPERATURE: Final[ControlValue] = ControlValue("TempCompressorIn.values.actValue", float)
+    COMPRESSOR_OUTPUT_TEMPERATURE: Final[ControlValue] = ControlValue("TempCompressorOut.values.actValue", float)
+    COMPRESSOR: Final[ControlValue] = ControlValue("Compressor.values.setValueScaled", float)
+    HIGH_PRESSURE: Final[ControlValue] = ControlValue("HighPressure.values.actValue", float)
+    LOW_PRESSURE: Final[ControlValue] = ControlValue("LowPressure.values.actValue", float)
 
 
 class HeatCircuit(Enum):
-    NAME: Final[ControlValue] = ControlValue("param.name", str, read_only=True)
-    TEMPERATURE: Final[ControlValue] = ControlValue("values.setValue", float, read_only=True)
+    NAME: Final[ControlValue] = ControlValue("param.name", str)
+    TEMPERATURE: Final[ControlValue] = ControlValue("values.setValue", float)
     DAY_TEMPERATURE: Final[ControlValue] = ControlValue("param.normalSetTemp", float, read_only=False)
-    DAY_TEMPERATURE_THRESHOLD: Final[ControlValue] = ControlValue("param.thresholdDayTemp.value", float, read_only=True)
+    DAY_TEMPERATURE_THRESHOLD: Final[ControlValue] = ControlValue("param.thresholdDayTemp.value", float)
     NIGHT_TEMPERATURE: Final[ControlValue] = ControlValue("param.reducedSetTemp", float, read_only=False)
-    NIGHT_TEMPERATURE_THRESHOLD: Final[ControlValue] = ControlValue(
-        "param.thresholdNightTemp.value",
-        float,
-        read_only=True,
-    )
-    HOLIDAY_TEMPERATURE: Final[ControlValue] = ControlValue("param.holidaySetTemp", float, read_only=True)
+    NIGHT_TEMPERATURE_THRESHOLD: Final[ControlValue] = ControlValue("param.thresholdNightTemp.value", float)
+    HOLIDAY_TEMPERATURE: Final[ControlValue] = ControlValue("param.holidaySetTemp", float)
     OFFSET_TEMPERATURE: Final[ControlValue] = ControlValue("param.offsetRoomTemp", float, read_only=False)
-    OPERATING_MODE: Final[ControlValue] = ControlValue("param.operatingMode", int, read_only=False)
+    OPERATING_MODE: Final[ControlValue] = ControlValue(
+        "param.operatingMode",
+        int,
+        read_only=False,
+        human_readable=HeatCircuitOperatingMode,
+    )
 
 
 Control: TypeAlias = Options | Outdoor | HotWaterTank | HeatPump | HeatCircuit